% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterJoinSTOICH.R
\name{filterJoinSTOICH}
\alias{filterJoinSTOICH}
\title{Filter Based on Simulated Joins of STOICH Data Tables}
\usage{
filterJoinSTOICH(dataTables, filtTable)
}
\arguments{
\item{dataTables}{The STOICH data tables filtered using filterSTOICH.}

\item{filtTable}{The table name filtering was performed on.}
}
\value{
Tables with corresponding entries that were removed from one table removed from the other tables.
}
\description{
Filters STOICH data tables following a filtering operation against rows of other tables to remove
any corresponding entries from other tables.
}
\examples{
\dontrun{
# Not typically run independently. This function is called by filterSTOICH
# to clean up before returning data.
}

}
\author{
Chad Petersen \email{cpetersen4@unl.edu}
}
