% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSTOICH.R
\name{loadSTOICH}
\alias{loadSTOICH}
\title{Load STOICH Data}
\usage{
loadSTOICH(dataPath = file.path(path.expand("~"), "data"))
}
\arguments{
\item{dataPath}{The path to the STOICH data directory.}
}
\value{
A list containing the STOICH data tables
}
\description{
Loads STOICH data from release files.
}
\examples{
\dontrun{
stoichData <- loadSTOICH(dataPath="C:/Users/example_user/Documents/data")
}

}
\author{
Chad Petersen \email{cpetersen4@unl.edu}
}
