% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifySTOICH.R
\name{verifySTOICH}
\alias{verifySTOICH}
\title{Check STOICH Data Table Format}
\usage{
verifySTOICH(dataTables)
}
\arguments{
\item{dataTables}{The STOICH data tables loaded using loadSTOICH.}
}
\value{
True if the STOICH data tables are the correct format or stop and return the first error found if an
incorrect format is detected.
}
\description{
Checks if the STOICH data tables loaded from release files are in the proper format for joining and filtering.
}
\examples{
\dontrun{
# Load the data
stoichData <- loadSTOICH(dataPath="C:/Users/example_user/Documents/data")

verifySTOICH(stoichData)

}

}
\author{
Chad Petersen \email{cpetersen4@unl.edu}
}
