% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoichUtilities-package.R
\docType{package}
\name{stoichUtilities-package}
\alias{stoichUtilities}
\alias{stoichUtilities-package}
\title{stoichUtilities: Provides user oriented tools to access STOICH Project data}
\description{
The stoichUtilities R package provides user tools for working with STOICH database files. This package is designed to enable easier and more robust access to the data files from STOICH database releases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/STOICH-project/STOICH-utilities}
  \item \url{https://snr-stoich.unl.edu/}
  \item \url{https://stoichproject.wordpress.com/}
  \item Report bugs at \url{https://github.com/STOICH-project/STOICH-utilities/issues}
}

}
\author{
\strong{Maintainer}: Chad Petersen \email{cpetersen4@unl.edu}

Other contributors:
\itemize{
  \item Jessica Corman \email{jcorman3@unl.edu} [contributor]
  \item Halvor Halvorson \email{hhalvorson@uca.edu} [contributor]
  \item Casey Brucker \email{cbrucker@uwyo.edu} [contributor]
  \item Eli Wess \email{ewess@uca.edu} [contributor]
  \item U.S. National Science Foundation [funder]
}

}
\keyword{internal}
