% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoichUtilities-package.R
\docType{package}
\name{stoichUtilities-package}
\alias{stoichUtilities}
\alias{stoichUtilities-package}
\title{stoichUtilities: User Tools for Accessing the STOICH Project Database}
\description{
User tools for working with The STOICH (Stoichiometric Traits of Organisms in their Chemical Habitats) Project database \url{https://snr-stoich.unl.edu/}. This package is designed to aid in data discovery, filtering, pairing water samples with organism samples, and merging data tables to assist users in preparing data for analyses. For additional examples see "Additional Examples" and the readme file at \url{https://github.com/STOICH-project/STOICH-utilities}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/STOICH-project/STOICH-utilities}
  \item \url{https://snr-stoich.unl.edu/}
  \item \url{https://stoichproject.wordpress.com/}
  \item Report bugs at \url{https://github.com/STOICH-project/STOICH-utilities/issues}
}

}
\author{
\strong{Maintainer}: Chad Petersen \email{cpetersen4@unl.edu}

Other contributors:
\itemize{
  \item Jessica Corman \email{jcorman3@unl.edu} [contributor]
  \item Halvor Halvorson \email{hhalvorson@uca.edu} [contributor]
  \item Casey Brucker \email{cbrucker@uwyo.edu} [contributor]
  \item Eli Wess \email{ewess@uca.edu} [contributor]
  \item U.S. National Science Foundation [funder]
}

}
\keyword{internal}
